// src/@core/purchase-data-types.ts
type YesOrNo = 'Y' | 'N'
// ...
export type CreatePurchaseOrderRequest = {
  number: string
  date: string
  due_date: string
  term: string
  ref?: string
  branch_id: number
  warehouse_id: number | undefined
  vendor_id: number | undefined
  total_qty_request: number
  total_qty_received?: number | 0
  total_qty_over?: number | 0
  total_discount_additional?: number | 0
  total_discount?: number | 0
  total_price: number
  total_tax?: number | 0
  total_charge?: number | 0
  grand_total?: number | 0
  description?: string | null
  created_at?: Date
  created_by?: string
  updated_at?: Date
  updated_by?: string
  archived?: YesOrNo
}
// ...
export type PurchaseOrderDetailResponse = {
  number: string
  sequence: number
  date: Date
  product_id: number
  unit_id: number
  branch_id: number
  qty_request: number
  qty_received: number
  qty_over: number
  price: number
  price_discount: number
  percent_discount: number
  price_tax: number
  percent_tax: number
  charge: number
  created_at?: Date | null
  created_by?: string | null
  updated_at?: Date | null
  updated_by?: string | null
  initial: string
  unit_name: string
  product_name: string
}